Mise à jour doc

+++ core/include/tsd/divers.hpp
@@ -69,8 +69,14 @@ extern float sinc(float T, float f);
 extern float sinc(float t);
 
 
-/** @brief Noyau de Dirichlet (à documenter !) */
-extern float Dirichlet(int N, float Ω);
+/** @brief Noyau de Dirichlet (à documenter !)
+ *
+ *  @f[
+ *  D_N(\Omega) = \sum_{k=-N}^N e^{-\mathbf{i}\Omega k}
+ *  @f]
+ *
+ * */
+extern double Dirichlet(int N, double Ω);
 
 /** @brief Polynôme de Chebychev du premier type : @f$T_n\left(\cos(\theta)\right) = \cos(n\theta)@f$ */
 extern Poly<float> Chebychev_T(int n);
